refactor: fee relayer and etc
