+++ packages/core/src/contexts/solana/constants.ts
@@ -1,5 +1,5 @@
 export const NETWORK_CONFIGS = {
   'mainnet-beta': {
-    endpoint: 'https://p2p.rpcpool.com/',
+    endpoint: `https://p2p.rpcpool.com/${process.env.REACT_APP_RPCPOOL_API_KEY}`,
   },
 };
