+++ packages/web/src/new/sdk/FeeRelayer/models/index.ts
@@ -2,3 +2,4 @@ export * from './FeeRelayer.StatsInfo';
 export * from './FeeRelayerError';
 export * from './FeeRelayerRequestType';
 export * from './TokenAccount';
+export * from './UsageStatus';
