--- packages/web/src/new/sdk/FeeRelayer/relay/RelayProgram/index.ts
@@ -1,1 +1,0 @@
-export * from './FeeRelayerRelayProgram';
