+++ packages/web/src/new/sdk/FeeRelayer/relay/index.ts
@@ -1,3 +1,6 @@
-export * from './FeeRelayerRelay';
-export * from './FeeRelayerRelayTopUp';
+export * from './FeeRelayer';
+export * from './FeeRelayer';
+export * from './FeeRelayerCalculator';
+export * from './FeeRelayerConstants';
+export * from './FeeRelayerContext';
 export * from './helpers';
