--- packages/web/src/new/sdk/FeeRelayer/reward/index.ts
@@ -1,1 +1,0 @@
-export * from './FeeRelayerRewardModels';
