+++ packages/web/src/new/sdk/FeeRelayer/swap/index.ts
@@ -1,1 +1,4 @@
+export * from './calculator';
 export * from './helper';
+export * from './SwapFeeRelayer';
+export * from './transactionBuilder';
