+++ packages/web/src/new/ui/components/common/ChooseWallet/CollectionView/index.ts
